fix(backend): split ratio fields
Fix an error with the outgoing payment `FixedDelivery` test.

The stream was delivering slightly too much money in this one case. It initially appeared to be a timing issue, brought about by the `await this.withdrawalService.create` in `MockPlugin`'s `sendData` implementation. But the underlying problem was that `@interledger/pay` expects the rate upper bound to be exclusive. The Rafiki backend was converting the rate to a float and then back to a ratio (when serialized to the database). This slight loss of precision resulted in `lowRate === highRate`, when it should always be `lowRate < highRate` (even if only slightly, e.g. `0.5 < 0.500000000001`).

The timing component of the bug was just quirk of how the packets are scheduled & computed. It isn't a bug in `@interledger/pay`.