chore(backend): add connector and account service (#132)
* chore(interledger): add balance service

* chore(interledger): add credit service

* chore(interledger): add asset service

* chore(interledger): add deposit service

* chore(interledger): add withdrawal service

* chore(interledger): add transfer service

* chore(interledger): add services test helper

* chore(interledger): add token service

* chore(interledger): get liquidity/settlement balance in asset service

* chore(interledger): service-ify accounts service

Restrict accounts table access to accounts service

* chore(interledger): don't lint, build, or test workspace

* chore(backend): add HTTP token service

* chore(backend): add balance service

* chore(backend): add transfer service

* chore(backend): add asset service

* chore(backend): add account service

* chore(backend): add deposit service

* chore(backend): add withdrawal service

* chore(backend): add credit service

* chore(backend): handle account transfers in account service

* chore(backend): remove connector graphql client service

* chore(backend): add connector

* chore(interledger): remove interledger package/workspace

* fix(backend): don't use input types for queries

* chore(backend): add errors and tests to transfer service

Specify timeout for two-phase commit transfers.
Move error enums to error files.

* chore(backend): replace non-null assertions with assert.ok

* fix(backend): misc review changes

Change default TigerBeetle port.
Add comments to transferFunds.
Fix logger formatting.
Log error.

* chore(backend): change string amounts to UInt64 custom scalar

* fix(backend): fix logging

Add logger to Apollo context.

* chore(backend): remove unused logger and ccp/routing

Don't exclude tests from compilation.

* fix(backend): split ratio fields

Fix an error with the outgoing payment `FixedDelivery` test.

The stream was delivering slightly too much money in this one case. It initially appeared to be a timing issue, brought about by the `await this.withdrawalService.create` in `MockPlugin`'s `sendData` implementation. But the underlying problem was that `@interledger/pay` expects the rate upper bound to be exclusive. The Rafiki backend was converting the rate to a float and then back to a ratio (when serialized to the database). This slight loss of precision resulted in `lowRate === highRate`, when it should always be `lowRate < highRate` (even if only slightly, e.g. `0.5 < 0.500000000001`).

The timing component of the bug was just quirk of how the packets are scheduled & computed. It isn't a bug in `@interledger/pay`.

Co-authored-by: sentientwaffle <sentientwaffle@gmail.com>