feat(backend): add docker build
Adding the ability to build a docker container for the backend. The build process is using the yarn
docker-build plugin. This uses `yarn pack` to add the code to the docker container. This is why the
`prepack` command has be added to ensure the code is compiled before being added to docker.

The `.npmignore` is required as by default `yarn pack` will use the `.gitignore` file if not specified.
Which results in the `dist` folder not being included in the pack.