feat(backend): add docker build to backend (#145)
Add the yarn workspace plugin into the repo. This allows commands such as
`yarn workspaces foreach` to be used to easily iterate over the various packages.

* feat(backend): build scripts

Getting the backend package to be able to be built locally within the package.
A clean step is added to remove the dist file and a copy-files step is added
to move across the schema file into the correct dist location. Typescript
is also added as a local dependency

* feat(yarn): docker build plugin

Adding a yarn docker build plugin - https://github.com/Dcard/yarn-plugins/tree/master/packages/docker-build
Adds functionality to yarn2 for easier docker builds by first copies everything needed for `yarn install` to `manifests` folder. It also runs `yarn pack` in all workspace dependencies and copies them to `packs` folder.

* chore(husky): dependently run

Check if the .husky file exists and only if it does run install. This prevents husky creating an error when used in docker containers.

* feat(backend): add docker build

Adding the ability to build a docker container for the backend. The build process is using the yarn
docker-build plugin. This uses `yarn pack` to add the code to the docker container. This is why the
`prepack` command has be added to ensure the code is compiled before being added to docker.

The `.npmignore` is required as by default `yarn pack` will use the `.gitignore` file if not specified.
Which results in the `dist` folder not being included in the pack.

* chore(pnp): fix deps