chore(backend): send/receive payments from/to settlement account
Outgoing payments and invoices have accounts for tracking the amounts
sent and received, but they do not represent (withdraw-able) balances.

Add sendOutgoingPayment mutation resolver. Wallet operators will be
notified when quote is complete (TODO) and call sendOutgoingPayment
after payment is approved (if necessary) and funds are reserved in
the sender's wallet account.