feat(incoming-payment): option to approve incoming payment
This commit introduces possibility, to approve incoming payment through Admin API. API can be called
with existing incoming payment as ID. Rafiki will fetch the incoming payment and update it's
approvedAt field in the database. In case that payment does not exists or that it is not in the
PENDING state, appropriate error is returned.

#2811