feat(incoming-payment): possibility to cancel incoming payment
It will be possible to cancel incoming payment due to some requirements. Rafiki just need to get a
call to cancel incoming payment, with the payment ID. If transaction is in PENDING state and it is
in the database, then it will get updated. It's cancelledAt field will be set  to current time.

#2811