docs: accts, idp, and disclosure component
Updated the accounting concepts page. Formatted the examples to be a bit easier to scan through. I'm working on revised examples outside of this PR. Also made some updates to the IDP doc and removed the use of the Disclosure component from the Webhook Events page and README. We are no longer using the component.