feat(frontend): POC for adding tenantId from session to headers for all apollo requests
Uses the assets and list asset query.
This POC passes the request to the listAsset function.
Which imports the apolloClient directly and passes
the cookie from request headers in the context.
To avoid having to set this on each query as we compose
it, my intention is to create a new getApolloClient
function and use that insteadof directly importing a
single client. This enables us to form a link to handle setting the headers per request (as opposed to static
links that are used across all requests as it is currently).