feat(fronted): tenanted admin api credentials (#3213)
* feat(frontend): set api credentials on session

* chore(frontend): more details in todo comment

* refactor(frontend): move credentials form from modal to component on index

* chore(frontend): mark dialog for removal

- not removing yet because not sure if we might end up using it.
could be useful if we want to make global redirect if this is not set.

* feat(frontend): store api creds in server side session

* feat(frontend): POC for adding tenantId from session to headers for all apollo requests

Uses the assets and list asset query.
This POC passes the request to the listAsset function.
Which imports the apolloClient directly and passes
the cookie from request headers in the context.
To avoid having to set this on each query as we compose
it, my intention is to create a new getApolloClient
function and use that insteadof directly importing a
single client. This enables us to form a link to handle setting the headers per request (as opposed to static
links that are used across all requests as it is currently).

* feat(frontend): form apollo client per request

- enables authLink to get tenantId, apiSecret from cookie in request
- wondered if this was a performance concern
(maybe why we had single instance before?) but found
several things indicating this is OK and even recommended:
    - https://github.com/apollographql/apollo-client/issues/9520#issuecomment-1067397043
    - https://www.apollographql.com/blog/how-to-use-apollo-client-with-remix

* fix(mock-ase): update seed script to pass tenant sig/id verifcation

* feat(frontend): block api cred form submit on invalid uuid

* feat(frontend): handle errors, WIP apollo client

- see TODOs in apollo client in frontend.
maybe need to remove some env vars and
verify how no tenantid/secret are handled

* feat(frontend): disable nav links

* docs(localenv): update readme to not say kratos is required

* chore(frontend): format

* chore(frontend): rm unused component

* chore(frontend): rm commented out code

* chore(frontend): formatting

* refactor(frontend): better error parsing

* chore(frontend): rm todo

* refactor(frontend): use session api for deletion, not manual

* fix(frontend): display error based on message

reverses previous commit to use apollo error. proved unreliable

* fix(frontend): rm SIGNATURE_SECRET, SIGNATURE_VERSION env vars

* feat(mock-ase): log operator/tenant details to streamline use of frontend

* feat(frontend): dont show nav items if api creds required and not set

* feat(frontend): move api credential set action to own endpoint

- removes the action from the index. the intention is to
expose the remix server port over docker and  call this
from the MASE to set the api credentials on start

* feat(frontend): prefill api credential form

* chore(frontend): format

* feat(frontend): auto submit form if values passed in

requires changing intent to be set as an input. submitting form bypasses the button so the action didnt have the intent  and
failed when auto submitting.

* fix: reinstate sig version env var