docs: accts, idp, and disclosure component (#3102)
* docs: accts, idp, and disclosure component

Updated the accounting concepts page. Formatted the examples to be a bit easier to scan through. I'm working on revised examples outside of this PR. Also made some updates to the IDP doc and removed the use of the Disclosure component from the Webhook Events page and README. We are no longer using the component.

* docs: IDP updates based on Max's feedback

* docs: idp and auth var updates

* docs: minor idp-related updates

* docs: updates to accounting doc

Running prettier caused updates to a bunch of other files. The only file I edited in this iteration was the accounting file.

* docs: updates based on feedback

Also fixed a table in the telemetry doc

* docs: feedback updates

* docs: fix broken link

* docs: prettier demands sacrifices

* docs: update accounting.mdx

* docs: slight wording update